Change rank datatype to int (#746)
Motivation: remove unnecessary divergence with libraries from the same domain (NCCL, NVSHEM, XCCL, MPI, Horovod, PyTorch). And "int" datatype has enough bits space for current and future cases.

+++ examples/benchmark/include/sycl_coll.hpp
@@ -16,7 +16,7 @@ using namespace sycl;
 using namespace sycl::access;
 
 std::vector<cl::sycl::queue> get_sycl_queues(sycl_dev_type_t dev_type,
-                                             const std::vector<size_t>& ranks) {
+                                             const std::vector<int>& ranks) {
 
     auto queues = create_sycl_queues(sycl_dev_names[dev_type], ranks);
