+++ examples/benchmark/src/alltoallv/sycl_alltoallv_coll.hpp
@@ -27,7 +27,7 @@ struct sycl_alltoallv_coll : sycl_base_coll<Dtype, alltoallv_strategy_impl> {
         if (base_coll::get_sycl_mem_type() != SYCL_MEM_BUF)
             return;
 
-        size_t local_rank = comm.rank();
+        int local_rank = comm.rank();
         for (size_t b_idx = 0; b_idx < base_coll::get_buf_count(); b_idx++) {
             stream.get_native().submit([&](handler& h) {
                 auto send_buf = (static_cast<sycl_buffer_t<Dtype>*>(send_bufs[b_idx][rank_idx]));
@@ -58,7 +58,7 @@ struct sycl_alltoallv_coll : sycl_base_coll<Dtype, alltoallv_strategy_impl> {
         }
 
         Dtype sbuf_expected = comm.rank();
-        size_t comm_size = comm.size();
+        int comm_size = comm.size();
 
         for (size_t b_idx = 0; b_idx < base_coll::get_buf_count(); b_idx++) {
             stream.get_native().submit([&](handler& h) {
