+++ include/oneapi/ccl/ccl_device.hpp
@@ -102,7 +102,7 @@ template<class DeviceType>
 using rank_device_pair_t = ccl::pair_class<size_t, typename std::remove_reference<typename std::remove_cv<DeviceType>::type>::type>;
 
 template <class device_value_type>
-constexpr auto attr_val(size_t rank, device_value_type&& v)
+constexpr auto attr_val(int rank, device_value_type&& v)
     -> rank_device_pair_t<device_value_type>{
     return rank_device_pair_t<device_value_type>{rank, std::forward<device_value_type>(v)};
 }
